refactor(geo): extract pure-domain Location, drop cache singleton
S3 / #128. Fixes F2 (geo transitively pulled [aws]) and the
Location._cache slice of F9 (class-level mutable singleton state).

- New hyperion/domain/geo.py: Location, NamedLocation, SpatialKMeans,
  meters_to_degrees, EARTH_RADIUS_METERS, LATITUDE_DEGREE_TO_METERS.
  Pure data + haversine math, no cache, no singletons.
- Removed Location._cache ClassVar and Location._get_cache(); get_distance
  no longer touches a cache. This is provably zero behaviour change: the old
  code only ever called cache.get() and never cache.set(), so the cache was
  always empty and the math path always ran. No application-layer caching
  service is needed (plan:238 conditional does not trigger).
- hyperion/infrastructure/geo/location.py is now a deferred-import
  deprecation shim (S2 _compat.moved_attr + PEP 562 __getattr__ recipe);
  every old symbol still resolves to the same object and emits a
  DeprecationWarning pointing at hyperion.domain.geo. Removed in 2.0.
- Migrated first-party imports (infrastructure/geo/__init__.py, gmaps.py)
  to hyperion.domain.geo; tests keep the old path as the deprecation signal.
- Import-graph guards: hyperion.domain.geo pulls no boto3 / data stack and
  not hyperion.ports.cache.

numpy decision (issue asked to decide & document): numpy CAN be [data]-gated.
SpatialKMeans imports numpy lazily so hyperion.domain.geo never requires it.
`haversine` does its own optional `import numpy` only when numpy is installed
(a vector fast-path, not a declared dependency) and falls back to a pure-math
scalar path otherwise -- the path Location.get_distance uses -- so a lite
install without numpy still computes distances; SpatialKMeans raises a clear
ImportError. The numpy -> [data] pyproject move itself is deferred to S10
(#136), the single build! step, keeping this a non-bumping refactor commit.

Closes #128
