+++ hyperion/infrastructure/geo/__init__.py
@@ -1,5 +1,5 @@
+from hyperion.domain.geo import Location
 from hyperion.infrastructure.geo.gmaps import GoogleMaps
-from hyperion.infrastructure.geo.location import Location
 
 __all__ = [
     "GoogleMaps",
