+++ tests/test_import_graph.py
@@ -30,11 +30,7 @@ HEAVY_MODULES = {
 
 def _loaded_modules(module_name: str) -> set[str]:
     """Spawn a fresh interpreter, import the module, return `sys.modules.keys()`."""
-    program = (
-        f"import {module_name}\n"
-        "import sys, json\n"
-        "print(json.dumps(list(sys.modules.keys())))\n"
-    )
+    program = f"import {module_name}\nimport sys, json\nprint(json.dumps(list(sys.modules.keys())))\n"
     result = subprocess.run(  # noqa: S603 - trusted args (sys.executable + literal program)
         [sys.executable, "-c", program],
         capture_output=True,
@@ -58,6 +54,7 @@ def _heavy_pulled_in(module_name: str) -> set[str]:
 
 # -- These modules must remain lite today and forever --
 
+
 @pytest.mark.parametrize(
     "module",
     [
@@ -78,6 +75,7 @@ def test_lite_core_module_pulls_no_heavy_deps(module: str) -> None:
 #    stack -- this guards the deferred-import / TYPE_CHECKING wiring. The cache
 #    and keyval ports legitimately use snappy for (de)compression, so snappy is
 #    deliberately not asserted against. --
+
 
 @pytest.mark.parametrize(
     "module",
@@ -107,6 +105,7 @@ def test_annotation_only_ports_stay_lite(module: str) -> None:
 #    and the entities.catalog / domain.assets identity layer no longer pulls
 #    the data stack. These must stay lite today and forever. --
 
+
 def test_typeutils_does_not_pull_data_stack() -> None:
     assert _heavy_pulled_in("hyperion.typeutils") == set()
 
@@ -119,9 +118,32 @@ def test_domain_assets_does_not_pull_data_stack() -> None:
     assert _heavy_pulled_in("hyperion.domain.assets") == set()
 
 
+# -- S3 (F2) landed: hyperion.domain.geo is pure data + haversine math. It must
+#    not pull boto3 / the data stack, and -- crucially for F2 -- must not pull
+#    the cache port (the Location._cache singleton that reached boto3 via
+#    Cache.from_config is gone). `numpy` is deliberately NOT asserted against:
+#    `haversine` (lite core) does an optional `import numpy` at module load
+#    *iff numpy is installed* (its vector fast-path); it is not a declared
+#    haversine dependency and the scalar path Location uses works without it, so
+#    SpatialKMeans imports numpy lazily and numpy is [data]-gated at the install
+#    level (Step 10 / the no-extras smoke test), not by this in-env guard. --
+
+
+def test_domain_geo_stays_lite() -> None:
+    forbidden = {"boto3", "botocore", "aioboto3", "polars", "pandera", "fastavro", "googlemaps"} & _heavy_pulled_in(
+        "hyperion.domain.geo"
+    )
+    assert forbidden == set(), f"hyperion.domain.geo unexpectedly imports {sorted(forbidden)}"
+
+
+def test_domain_geo_does_not_pull_cache_port() -> None:
+    assert "hyperion.ports.cache" not in _loaded_modules("hyperion.domain.geo")
+
+
 # -- This module currently violates the promise; the refactor fixes it --
 # `strict=True` means the marker fails CI as soon as the test starts passing —
 # forcing the marker to be removed when the corresponding refactor step lands.
+
 
 @pytest.mark.xfail(
     strict=True,
